Fix a bug for file source

+++ app/src/main/java/com/mantz_it/rfanalyzer/MainActivity.java
@@ -600,7 +600,7 @@ public class MainActivity extends Activity implements IQSourceInterface.Callback
 			case FILE_SOURCE:
 						// Create IQ Source (filesource)
 						try {
-							frequency = Integer.valueOf(preferences.getString(getString(R.string.pref_filesource_frequency), "97000000"));
+							frequency = Long.valueOf(preferences.getString(getString(R.string.pref_filesource_frequency), "97000000"));
 							sampleRate = Integer.valueOf(preferences.getString(getString(R.string.pref_filesource_sampleRate), "2000000"));
 						} catch (NumberFormatException e) {
 							this.runOnUiThread(new Runnable() {
